Improved printing support (#380)
This commit improves the printing support and fixes #352 and fixes #320.
The following list summarizes the current printing behavior:

* Reference types `T` are printed as `T@0x012345`, where the hexadecimal
  value is the address of the object. The same thing goes for futures,
  streams, arrays, ParTs and closures. `null` is printed as `T@0x0`.

* Values whose type isn't statically known cannot be printed. This
  includes any polymorphic values.

* Values of embedded types cannot be printed. You can (and should) drop
  down to C if you really need to print such a value (or closures or
  polymorphic values for that matter).

* Ranges are printed as `[1..10 by 1]`. The `by` part is always included
  for simplicity.

* Maybe types are printed like they should, although the generated code
  is a bit nasty. For each maybe type, a string is allocated for the
  sole purpose of being printed. The length of this string is
  approximated by summing approximations of the lengths of all
  subexpressions. Strings use their dynamic lengths and tuples the sum
  of the approximations of their contents. All other lengths are
  approximated as the length of their type plus 30 (compare with how
  reference types are printed). The 30 should be enough to hold any
  addresses, integers or reals. It would be better to calculate a
  maximum depending on the type, but if you're doing things like
  printing `Maybe`s, you probably don't care about performance anyway.

* Tuples are also printed like they should, and since their shape is
  known statically, the code isn't as bad.

* Values of `void` type are always printed as `()`

* Strings, chars, ints, booleans and reals are printed like you would
  expect. The only news here is proper printing of booleans as `true`
  and `false`.

The existing tests already test much of the printing, and some of them
have been updated to align with the new syntax. The test `printf` tests
some of the new printing.

Testing this has reminded me that the lack of proper inference for
`null` and `bottom` is annoying, but fixing this does not belong in this
PR.

+++ src/runtime/range/range.c
@@ -5,8 +5,8 @@
 
 struct range_t
 {
-  size_t start;
-  size_t stop;
+  int64_t start;
+  int64_t stop;
   int64_t step;
 };
 
@@ -22,7 +22,8 @@ void range_trace(pony_ctx_t* ctx, void *p)
   // No pointers
 }
 
-struct range_t *range_mk(pony_ctx_t* ctx, size_t start, size_t stop, int64_t step)
+struct range_t *range_mk(pony_ctx_t* ctx, int64_t start,
+                         int64_t stop, int64_t step)
 {
   ctx = pony_ctx();
   range_assert_step(step);
@@ -40,6 +41,6 @@ void range_assert_step(int64_t step)
     }
 }
 
-size_t  range_start (struct range_t *a) { return a->start; }
-size_t  range_stop  (struct range_t *a) { return a->stop;  }
+int64_t  range_start (struct range_t *a) { return a->start; }
+int64_t  range_stop  (struct range_t *a) { return a->stop;  }
 int64_t range_step  (struct range_t *a) { return a->step;  }
